requirements.txt
aiohttp==3.9.4
aiosignal==1.3.1
aiostream==0.6.1
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
asttokens
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
comm
contourpy==1.3.0
cycler==0.12.1
dashscope==1.20.0
dataclasses-json==0.6.4
debugpy
decorator
Deprecated==1.2.14
distro==1.9.0
exceptiongroup
executing
fastapi==0.110.2
filelock==3.13.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.3.1
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel 
ipython 
jedi
jsonpatch==1.33
jsonpointer==2.4
jupyter_client
jupyter_core
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.17
langchain-chroma==0.1.0
langchain-community==0.0.36
langchain-core==0.1.50
langchain-openai==0.1.6
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langgraph==0.0.40
langsmith==0.1.45
loguru==0.7.2
lxml==5.2.2
markdown-it-py==3.0.0
marshmallow==3.21.1
matplotlib==3.9.2
matplotlib-inline
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest_asyncio
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.25.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
parso
pexpect
pickleshare
pillow==10.4.0
platformdirs
posthog==3.5.0
prompt-toolkit
protobuf==4.25.3
psutil 
ptyprocess
pulsar-client==3.5.0
pure-eval
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic_core==2.18.1
Pygments
pyparsing==3.1.4
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq
regex==2023.12.25
reportlab==4.2.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
stack-data
starlette==0.37.2
sympy==1.12
tavily-python==0.3.3
tenacity==8.2.3
tiktoken==0.6.0
tokenizers==0.19.1
tornado
tqdm==4.66.2
traitlets
typer==0.12.3
types-requests==2.31.0.20240406
typing-inspect==0.9.0
typing_extensions
tzlocal==5.2
urllib3==2.2.1
# not use in windows
# uvicorn==0.29.0
# uvloop==0.19.0
watchfiles==0.21.0
wcwidth
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp
